requirements.txt
asgiref==3.8.1
cffi==1.17.1
coverage==7.6.10
cryptography==44.0.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
